pyproject.toml
[project]
name = "adk-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk==1.21.0",
    "litellm==1.80.10",
    "psutil==7.1.3",
    "python-dotenv==1.2.1",
    "ruff>=0.14.10",
    "yfinance>=0.2.66",
]
